Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0

